fix sysview construction
