+++ ydb/core/tx/columnshard/engines/reader/sys_view/abstract/granule_view.cpp
@@ -1,0 +1,5 @@
+#include "granule_view.h"
+
+namespace NKikimr::NOlap::NReader::NSysView::NAbstract {
+
+}
