+++ ydb/core/tx/columnshard/engines/reader/sys_view/abstract/iterator.cpp
@@ -1,0 +1,5 @@
+#include "iterator.h"
+
+namespace NKikimr::NOlap::NReader::NSysView::NAbstract {
+
+}
