+++ ydb/core/tx/columnshard/engines/reader/sys_view/abstract/metadata.cpp
@@ -1,4 +1,4 @@
-#include "abstract.h"
+#include "metadata.h"
 
 namespace NKikimr::NOlap::NReader::NSysView::NAbstract {
